+++ examples/run_async.py
@@ -1,52 +1,50 @@
 #!/usr/bin/env python -u
 
+"""
+This example will execute 10 queries in total, 2 concurrent queries at a time.
+Each query will sleep for 2 seconds before returning.
+Here's a sample output that shows that the queries are executed concurrently in batches of 2:
+```
+Completed query 1, elapsed ms since start: 2002
+Completed query 0, elapsed ms since start: 2002
+Completed query 3, elapsed ms since start: 4004
+Completed query 2, elapsed ms since start: 4005
+Completed query 4, elapsed ms since start: 6006
+Completed query 5, elapsed ms since start: 6007
+Completed query 6, elapsed ms since start: 8009
+Completed query 7, elapsed ms since start: 8009
+Completed query 9, elapsed ms since start: 10011
+Completed query 8, elapsed ms since start: 10011
+```
+"""
+
 import asyncio
 from datetime import datetime
 
 import clickhouse_connect
 
-QUERIES = 100
-SEMAPHORE = 10
-
-clickhouse_connect.common.set_setting("autogenerate_session_id", False)
-
-
-class AsyncRunner:
-    def __init__(self, **client_params):
-        self.client = clickhouse_connect.create_client(**client_params)
-        self.query_count = 0
-        self.row_count = 0
-
-    async def run_query(self, query: str, num: int):
-        self.query_count += 1
-
-        def read_data():
-            return self.client.query(query).result_set
-
-        result_set = await asyncio.to_thread(read_data)
-        self.row_count += len(result_set)
-        print(f"completed query {num}")
-        return result_set
+QUERIES = 10
+SEMAPHORE = 2
 
 
 async def concurrent_queries():
-    test_query = ("SELECT * FROM generateRandom('a Array(Int8), d Decimal32(4)," +
-                  " c Tuple(DateTime64(3), UUID)', 1, 10, 2) LIMIT 200000")
-    runner = AsyncRunner()
+    test_query = "SELECT sleep(2)"
+    client = await clickhouse_connect.get_async_client()
 
-    print(datetime.now())
+    start = datetime.now()
 
     async def semaphore_wrapper(sm: asyncio.Semaphore, num: int):
         async with sm:
-            await runner.run_query(test_query, num)
+            await client.query(query=test_query)
+            print(f"Completed query {num}, "
+                  f"elapsed ms since start: {int((datetime.now() - start).total_seconds() * 1000)}")
 
     semaphore = asyncio.Semaphore(SEMAPHORE)
     await asyncio.gather(*[semaphore_wrapper(semaphore, num) for num in range(QUERIES)])
-    print(f"{datetime.now()}  query_count: {runner.query_count}  row_count: {runner.row_count}")
 
 
-loop = asyncio.get_event_loop()
-try:
-    loop.run_until_complete(concurrent_queries())
-finally:
-    loop.close()
+async def main():
+    await concurrent_queries()
+
+
+asyncio.run(main())